Fix missing behavior warnings, improve UI
The missing behavior warnings were do to the fact that hydrolysis by default
doesn't look outside of the base dir of a file. However, having every element
doing its own analysis of itself and all of its dependencies is crazy slow and
uses tons of memory.

The simple solution was to analyze all of the HTML for all elements in one
place before starting the cleaning process.

This commit also comes with a bunch of UI improvements:

- show progress while analyzing elements with hydrolysis
- don't push anything by default, added a CLI option to give the max number
  of elements to push up. very useful when developing and unsure
- improved reporting of what changes were pushed and remain to push
- standardized progress bar widths
- support installing tedium as a normal binary (npm install -g)