Track more metadata about cleanup passes.
Also moves the blacklist into a config file. My thinking here is
that we're going to want to expand this config file in the future
such that all of the PolymerElements-specific material will live
in the config file, making it possible for other folks to use
tedium to do README generation, etc.